refactor: consistent app-routes name
